Fix niche bug related to ECS configurable/promise return type
Promise and configurable data (including return type) was provided as an expression rather than the actual type. So if a child component overrode a promise or configurable of its parent, the parent's listed return type for that promise/configurable had to be a valid expression from the child's POV too. With this change, that is no longer necessary.